refactor report post module
